# Reduce allocations for queries on `HEAD`