This was growing the size of the `[]string` while iterating over, which in practive ends up being multiple allocations. Since we know the max number that could be there, we should instead right-size allocate the slice. In my local testing, here are the results (from the benchmark included):

```
$ benchcmp baseline after
benchmark                             old ns/op     new ns/op     delta
BenchmarkHeadPostingForMatchers-4     74452         63514         -14.69%

benchmark                             old allocs     new allocs     delta
BenchmarkHeadPostingForMatchers-4     20             13             -35.00%

benchmark                             old bytes     new bytes     delta
BenchmarkHeadPostingForMatchers-4     5425          3137          -42.18%
```